邮件和交换箱导入时，数据类型int换为long

+++ src/main/java/com/vrv/vap/apicasom/business/meeting/dao/impl/SituationLargeScreenDaoImpl.java
@@ -87,7 +87,7 @@ public class SituationLargeScreenDaoImpl implements SituationLargeScreenDao {
         public TreandVO mapRow(ResultSet rs, int rowNum) throws SQLException {
             TreandVO data = new TreandVO();
             data.setDataX(rs.getString("dataX"));
-            data.setDataY(rs.getInt("dataY"));
+            data.setDataY(rs.getLong("dataY"));
             return data;
         }
     }
